Done: customer ordering

+++ src/app/customer-cart-page/customer-cart-page.component.ts
@@ -3,6 +3,7 @@ import { CartService } from './../services/cart.service';
 import { Component, OnInit } from '@angular/core';
 import { ItemCart } from '../item/item.cart';
 import { UserBusiness } from '../user/userbusiness';
+import { UserCustomer } from '../user/usercustomer';
 
 @Component({
   selector: 'app-customer-cart-page',
@@ -11,24 +12,61 @@ import { UserBusiness } from '../user/userbusiness';
 })
 export class CustomerCartPageComponent implements OnInit {
 
-  constructor(private cartService: CartService, private authService: AuthService) { }
+  constructor(private cartService: CartService, private authService: AuthService) 
+  {
+  }
 
   carts: ItemCart[][];
   businessData: UserBusiness[];
+  finalPrice: number[];
+  confirmed: boolean[];
 
-  async ngOnInit() {
-    // await this.cartService.getCartsForUser(this.authService.userID);
+  ngOnInit() {
+    this.getData();
+  }
+
+  getData()
+  {
     this.carts = this.cartService.carts;
     this.businessData = this.cartService.businessData;
+    this.finalPrice = this.cartService.finalPrice;
   }
 
   public cartsAvailable(): boolean
-  {
+  { 
     return this.carts && this.carts[0] && this.carts[0].length > 0;
   }
 
-  public removeFromCart(numbers: any)
+  public async removeFromCart(numbers: any)
   {
-    this.cartService.removeItemFromCart(numbers.cartNum, numbers.itemNum);
+    if(this.getConfirmed(numbers.cartNum))
+      return;
+
+    await this.cartService.removeItemFromCart(numbers.cartNum, numbers.itemNum);
+    this.getData();
   }
+
+  public async onOrder(cartNum: number)
+  {
+    if(this.getConfirmed(cartNum))
+      return;
+
+    await this.cartService.confirmOrder(cartNum, this.authService.userData as UserCustomer);
+    this.confirmed = this.cartService.confirmed;
+  }
+
+  async confirmQuantity(item: ItemCart)
+  {
+    if(this.getConfirmed(cartNum))
+      return;
+      
+    await this.cartService.updateQuantity(item);
+    this.getData();
+  }
+
+  getConfirmed(i: number)
+  {
+    return this.confirmed && this.confirmed[i] != null && this.confirmed[i] != undefined && this.confirmed[i];
+  }
+  
 }
